Update pingvin.sh and pingvin-install.sh (#1819)

+++ ct/pingvin.sh
@@ -32,29 +32,41 @@ function update_script() {
         msg_error "No ${APP} Installation Found!"
         exit
     fi
-    msg_info "Stopping Pingvin Share"
-    systemctl stop pm2-root.service
-    msg_ok "Stopped Pingvin Share"
+   
+    RELEASE=$(curl -s https://api.github.com/repos/stonith404/pingvin-share/releases/latest | grep "tag_name" | awk '{print substr($2, 3, length($2)-4) }')
+    if [[ ! -f /opt/$pingvin_version.txt ]] || [[ "${RELEASE}" != "$(cat /opt/pingvin_version.txt)" ]]; then
+      
+      msg_info "Stopping Pingvin Share"
+      systemctl stop pm2-root.service
+      msg_ok "Stopped Pingvin Share"
+      
+      msg_info "Updating Pingvin Share to v${RELEASE}"
+      cd /opt
+      wget -q "https://github.com/stonith404/pingvin-share/archive/refs/tags/v${RELEASE}.zip"
+      unzip -q v${RELEASE}.zip
+      mv pingvin-share-${RELEASE} /opt/pingvin-share  
+      cd /opt/pingvin-share
+      cd backend
+      npm install &>/dev/null
+      npm run build &>/dev/null
+      cd ../frontend
+      npm install &>/dev/null
+      npm run build &>/dev/null
+      echo "${RELEASE}" >"/opt/pingvin_version.txt"
+      rm -rf /opt/v${RELEASE}.zip
+      msg_ok "Updated Pingvin Share to v${RELEASE}"
 
-    msg_info "Updating Pingvin Share"
-    cd /opt/pingvin-share
-    git fetch --tags
-    git checkout $(git describe --tags $(git rev-list --tags --max-count=1)) &>/dev/null
-    cd backend
-    npm install &>/dev/null
-    npm run build &>/dev/null
-    cd ../frontend
-    npm install &>/dev/null
-    npm run build &>/dev/null
-    msg_ok "Updated Pingvin Share"
+      msg_info "Starting Pingvin Share"
+      systemctl start pm2-root.service
+      msg_ok "Started Pingvin Share"
 
-    msg_info "Starting Pingvin Share"
-    systemctl start pm2-root.service
-    msg_ok "Started Pingvin Share"
-
-    msg_ok "Updated Successfully"
-    exit
+      msg_ok "Updated Successfully"
+      exit
+  else
+    msg_ok "No update required. Pingvin Share is already at v${RELEASE}."
+  fi
 }
+
 
 start
 build_container
